+++ code/core/src/main/java/org/betonquest/betonquest/quest/placeholder/translate/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.placeholder.NullablePlaceholder NullablePlaceholder}
+ * implementation for language values in constants.
+ */
+package org.betonquest.betonquest.quest.placeholder.translate;