+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - `language.questlang_whitelist` to restrict which languages players can select via `/questlang`
 - nesting in quest packages for single line instruction types
+- `translate` placeholder parsing languages from `constants` section
 ### Changed
 - `language` config option is now a section; the default language moved to `language.default`
 ### Deprecated